pyproject.toml
[project]
name = "pmagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.2",
    "django>=6.0",
    "django-environ>=0.12.0",
    "django-ninja>=1.5.1",
    "gunicorn[gevent]>=23.0.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "orjson>=3.11.5",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]
